Simplify the overriding of Trepn preferences
This commit extends the overriding logic of Trepn preferences into a self-contained function, allowing users to override all configurations. Mirror the preferences key in the user-defined configuration to extend the overriding to any preference. To reduce the verbosity of the key, the mirrored key contains only the substring after the last dot. For instance, the preference `com.quicinc.preferences.general.profiling_interval` is mapped into `profiling_interval`.